Adding default dimensions for SVG icons
Without such dimensions, they may show up in a huge size for a fraction
of a second, after the DOM has been constructed but before the CSS has
been applied by the browser. This can be the case if this component is
used in Nuxt.js, which renders the page server-side.

The width/height as attributes to <svg> element are overwritten by CSS
rules. (They already set `width: 1em; height: 1em;`.) Thus, the result
of this commit is only visible for a fraction of a second (only if this
component is rendered upon page load, and before the CSS applied).
Still, I believe this commit improves the UX.

The default width/height value is `auto`, which means `100%`, which
makes the icon as large as the available width.

https://developer.mozilla.org/en-US/docs/Web/SVG/Attribute/width#svg